Added routine to access components without specifying if wrapped
A lot of components in this Gem are created wrapped in
GenericComponentWrapper. Those need to be accessed differently in game
mode and editor. `GetGameOrEditorComponent` was added to help with that.
