+++ Code/Source/RobotImporter/Utils/RobotImporterUtils.cpp
@@ -5,9 +5,9 @@
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  *
  */
+
 #include "RobotImporterUtils.h"
 #include <AzCore/std/string/regex.h>
-#include <AzToolsFramework/API/EntityCompositionRequestBus.h>
 namespace ROS2
 {
 
@@ -38,31 +38,6 @@ namespace ROS2
             return true;
         }
         return false;
-    }
-
-    AZ::ComponentId Utils::CreateComponent(const AZ::EntityId entityId, const AZ::Uuid componentType)
-    {
-        AZ::ComponentTypeList componentsToAdd{ componentType };
-        AZStd::vector<AZ::EntityId> entityIds{ entityId };
-        AzToolsFramework::EntityCompositionRequests::AddComponentsOutcome addComponentsOutcome = AZ::Failure(AZStd::string());
-        AzToolsFramework::EntityCompositionRequestBus::BroadcastResult(
-            addComponentsOutcome, &AzToolsFramework::EntityCompositionRequests::AddComponentsToEntities, entityIds, componentsToAdd);
-        if (!addComponentsOutcome.IsSuccess())
-        {
-            AZ_Warning(
-                "URDF importer",
-                false,
-                "Failed to create component %s, entity %s : %s",
-                componentType.ToString<AZStd::string>().c_str(),
-                entityId.ToString().c_str(),
-                addComponentsOutcome.GetError().c_str());
-        }
-        const auto& added = addComponentsOutcome.GetValue().at(entityId).m_componentsAdded;
-        if (added.size())
-        {
-            return added.front()->GetId();
-        }
-        return AZ::InvalidComponentId;
     }
 
 } // namespace ROS2
